% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{cppirt_model}
\alias{cppirt_model}
\title{Fitting the two parameter logistic IRT model}
\usage{
cppirt_model(data, priors = list(), control = list())
}
\arguments{
\item{data}{an I (individual) x J (item) matrix. This should contain 0 or 1.}

\item{priors}{a list containing priors:
\itemize{
  \item \code{a0}: a double. Prior mean of alpha. Default is 0.
  \item \code{A0}: a double (> 0). Prior variance of alpha. Default is 25.
  \item \code{b0}: a double. Prior mean of beta. Default is 0.
  \item \code{B0}: a double (> 0). Prior variance of beta. Default is 25.
}}

\item{control}{a list containing model controls:
\itemize{
  \item \code{theta_constraint}: a positive integer. The position of an individual whose \code{theta} is always set positive.
  \item \code{theta_strict_identification}: a bool. Whether to standardize \code{theta} for identification.
  \item \code{maxit}: a positive integer. Maximum number of EM iteration. Default is 500.
  \item \code{verbose}: a positive integer. Print the status every \code{verbose} iteration. If this is not specified, any iteration status is not be printed.
  \item \code{tol}: a float (0.0 < tol < 1.0). Convergence threshold. Default is 1e-6.
}}
}
\value{
A \code{\link{CppIRTModel}} object.
}
\description{
\code{\link{cppirt_model}} is an instance to estimate IRT. For the methods, see \code{\link{CppIRTModel}}.
}
\examples{
\dontrun{
 library(CppIRT)
  
 # data generating process
 I <- 1000
 J <- 100
 alpha <- rnorm(J)
 beta <- rnorm(J)
 theta <- rnorm(I)
 p <- plogis(
  cbind(1, theta) \%*\% rbind(alpha, beta)
 )
 Y <- rbinom(I * J, 1, p) |>
        matrix(I, J)
  
 # create an instance
 model <- cppirt_model(
   data = Y,
   control = list(
     theta_constraint = which.max(theta),
     theta_strict_identification = TRUE
   )
 )
  
 # fit the model
 model$fit()
  
 # get the result
 out <- model$output()
}
}
