% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{CppIRTModel}
\alias{CppIRTModel}
\title{CppIRTModel objects}
\description{
A \code{\link{CppIRTModel}} object is an \code{\link[R6]{R6Class}} object created
  by the \code{\link{cppirt_model}} function.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CppIRTModel-new}{\code{CppIRTModel$new()}}
\item \href{#method-CppIRTModel-fit}{\code{CppIRTModel$fit()}}
\item \href{#method-CppIRTModel-output}{\code{CppIRTModel$output()}}
\item \href{#method-CppIRTModel-clone}{\code{CppIRTModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CppIRTModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-CppIRTModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the new instance of \code{\link{CppIRTModel}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppIRTModel$new(data, priors, control)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{an I (individual) x J (item) matrix. This should contain 0 or 1.}

\item{\code{priors}}{a list containing priors:
\itemize{
  \item \code{a0}: a double. Prior mean of alpha. Default is 0.
  \item \code{A0}: a double (> 0). Prior variance of alpha. Default is 25.
  \item \code{b0}: a double. Prior mean of beta. Default is 0.
  \item \code{B0}: a double (> 0). Prior variance of beta. Default is 25.
}}

\item{\code{control}}{a list containing model controls:
\itemize{
  \item\code{theta_constraint}: a positive integer. The position of an individual whose \code{theta} is always set positive.
  \item\code{theta_strict_identification}: a bool. Whether to standardize \code{theta} for identification.
  \item\code{maxit}: a positive integer. Maximum number of EM iteration. Default is 500.
  \item\code{verbose}: a positive integer. Print the status every \code{verbose} iteration. If this is not specified, any iteration status is not be printed.
  \item\code{tol}: a float (0.0 < tol < 1.0). Convergence threshold. Default is 1e-6.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CppIRTModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-CppIRTModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model via Polya-Gamma data augmentation and EM algorithm. Note that this method is a void function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppIRTModel$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CppIRTModel-output"></a>}}
\if{latex}{\out{\hypertarget{method-CppIRTModel-output}{}}}
\subsection{Method \code{output()}}{
Return the model fitting result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppIRTModel$output()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CppIRTModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CppIRTModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppIRTModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
